Cast hex values as bytes, as they are ints in Java. Fix typo in sprite 6.

+++ src/main/java/fortville/opcodes/SetIToSpriteLocation.java
@@ -15,6 +15,6 @@ public class SetIToSpriteLocation implements Opcode {
     public void execute(short data1, short data2, short data3,
         Memory memory, Display display, Registers registers) {
 
-        registers.storeI(5*registers.loadRegister(data1));
+        registers.storeI((short)(5 * registers.loadRegister(data1)));
     }
 }
